refactor(shim): PR-9.6 — 删 event_bus / log_tail / argparse_bridge / llm_presets / migrations 5 shim
batch 4 of 6 — top-level → infrastructure 5 个 shim。

  删 studio/{event_bus,log_tail,argparse_bridge,llm_presets,migrations}.py

caller 更新（~20 处）：
  studio/api/lifespan.py / events_sse.py / generate.py: from ..event_bus → ..infrastructure.event_bus
  studio/api/routers/{queue/io,lifecycle,outputs,projects/_shared,training,exports,curation}.py:
    from ....event_bus → ....infrastructure.event_bus
  studio/supervisor/{core,slot}.py: from ..log_tail → ..infrastructure.log_tail
  studio/services/models/downloader.py: from ...event_bus → ...infrastructure.event_bus
  runtime/training/{cli,bootstrap}.py: from studio.argparse_bridge → studio.infrastructure.argparse_bridge
  tests/test_{argparse_bridge,log_tail,event_bus_callbacks,migrations,version_status_phase}.py:
    更新 import 路径到 infrastructure 包

回归：1579 passed（baseline 1568；新增 11 个测试过 — 一些 collection error
被解开。argparse + imagehash + lycoris fail 全 pre-existing）。
