+++ studio/api/routers/queue/io.py
@@ -22,7 +22,7 @@ from ...schemas.queue import ImportRequest
 from .... import db
 from ....services.presets import io as presets_io
 from ....services import queue_io, task_snapshot
-from ....event_bus import bus
+from ....infrastructure.event_bus import bus
 
 router = APIRouter()
