+++ studio/api/routers/queue/outputs.py
@@ -26,7 +26,7 @@ from ...errors import _export_result, _safe_join_or_400, _unique_data_export_pat
 from ...schemas.queue import ExportOutputsBody
 from .... import db
 from ....services.projects import projects, versions
-from ....event_bus import bus
+from ....infrastructure.event_bus import bus
 from ....paths import DATA_EXPORTS
 
 router = APIRouter()
